requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==4.3.0
apify_client==1.6.4
apify_shared==1.1.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.1.0
backports.tarfile==1.0.0
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.2.2
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
docopt==0.6.2
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.4
FLAML==2.1.2
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
googlesearch-python==1.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.29.4
ipython==8.23.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==25.1.0
lxml==5.2.1
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.4
nest-asyncio==1.6.0
newspaper3k==0.2.8
nh3==0.2.17
nltk==3.8.1
numpy==1.26.1
openai==1.17.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.10.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyautogen==0.2.22
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
readme_renderer==43.0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.4.2
scipy==1.13.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.27.2
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.4.0
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.2
twine==5.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.13
yarl==1.9.2
zipp==3.17.0
